FAC-963:impl:Move assessmentinsight packages to insight package
