--- flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/assessmentinsight/CreateAssessmentInsightRequestDto.java
@@ -1,4 +1,0 @@
-package org.flickit.assessment.core.adapter.in.rest.assessmentinsight;
-
-public record CreateAssessmentInsightRequestDto(String insight) {
-}
