+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/insight/assessmentinsight/CreateAssessmentInsightRequestDto.java
@@ -1,0 +1,4 @@
+package org.flickit.assessment.core.adapter.in.rest.insight.assessmentinsight;
+
+public record CreateAssessmentInsightRequestDto(String insight) {
+}
