+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/insight/assessmentinsight/CreateAssessmentInsightRestController.java
@@ -1,9 +1,9 @@
-package org.flickit.assessment.core.adapter.in.rest.assessmentinsight;
+package org.flickit.assessment.core.adapter.in.rest.insight.assessmentinsight;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.config.jwt.UserContext;
-import org.flickit.assessment.core.application.port.in.assessmentinsight.CreateAssessmentInsightUseCase;
-import org.flickit.assessment.core.application.port.in.assessmentinsight.CreateAssessmentInsightUseCase.Param;
+import org.flickit.assessment.core.application.port.in.insight.assessmentinsight.CreateAssessmentInsightUseCase;
+import org.flickit.assessment.core.application.port.in.insight.assessmentinsight.CreateAssessmentInsightUseCase.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
