+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/insight/assessmentinsight/GetAssessmentInsightRestController.java
@@ -1,10 +1,10 @@
-package org.flickit.assessment.core.adapter.in.rest.assessmentinsight;
+package org.flickit.assessment.core.adapter.in.rest.insight.assessmentinsight;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.config.jwt.UserContext;
-import org.flickit.assessment.core.application.port.in.assessmentinsight.GetAssessmentInsightUseCase;
-import org.flickit.assessment.core.application.port.in.assessmentinsight.GetAssessmentInsightUseCase.Param;
-import org.flickit.assessment.core.application.port.in.assessmentinsight.GetAssessmentInsightUseCase.Result;
+import org.flickit.assessment.core.application.port.in.insight.assessmentinsight.GetAssessmentInsightUseCase;
+import org.flickit.assessment.core.application.port.in.insight.assessmentinsight.GetAssessmentInsightUseCase.Param;
+import org.flickit.assessment.core.application.port.in.insight.assessmentinsight.GetAssessmentInsightUseCase.Result;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
