+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/port/in/insight/assessmentinsight/GetAssessmentInsightUseCase.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.application.port.in.assessmentinsight;
+package org.flickit.assessment.core.application.port.in.insight.assessmentinsight;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
@@ -6,20 +6,21 @@ import lombok.EqualsAndHashCode;
 import lombok.Value;
 import org.flickit.assessment.common.application.SelfValidating;
 
+import java.time.LocalDateTime;
 import java.util.UUID;
 
 import static org.flickit.assessment.common.error.ErrorMessageKey.COMMON_CURRENT_USER_ID_NOT_NULL;
-import static org.flickit.assessment.core.common.ErrorMessageKey.INIT_ASSESSMENT_INSIGHT_ASSESSMENT_ID_NOT_NULL;
+import static org.flickit.assessment.core.common.ErrorMessageKey.GET_ASSESSMENT_INSIGHT_ASSESSMENT_ID_NOT_NULL;
 
-public interface InitAssessmentInsightUseCase {
+public interface GetAssessmentInsightUseCase {
 
-    void initAssessmentInsight(Param param);
+    Result getAssessmentInsight(Param param);
 
     @Value
     @EqualsAndHashCode(callSuper = false)
     class Param extends SelfValidating<Param> {
 
-        @NotNull(message = INIT_ASSESSMENT_INSIGHT_ASSESSMENT_ID_NOT_NULL)
+        @NotNull(message = GET_ASSESSMENT_INSIGHT_ASSESSMENT_ID_NOT_NULL)
         UUID assessmentId;
 
         @NotNull(message = COMMON_CURRENT_USER_ID_NOT_NULL)
@@ -30,6 +31,12 @@ public interface InitAssessmentInsightUseCase {
             this.assessmentId = assessmentId;
             this.currentUserId = currentUserId;
             this.validateSelf();
+        }
+    }
+
+    record Result(Insight defaultInsight, Insight assessorInsight, boolean editable, boolean approved) {
+
+        public record Insight(String insight, LocalDateTime creationTime, boolean isValid) {
         }
     }
 }
