+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/service/insight/assessmentinsight/ApproveAssessmentInsightService.java
@@ -1,9 +1,9 @@
-package org.flickit.assessment.core.application.service.assessmentinsight;
+package org.flickit.assessment.core.application.service.insight.assessmentinsight;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.application.domain.assessment.AssessmentAccessChecker;
 import org.flickit.assessment.common.exception.AccessDeniedException;
-import org.flickit.assessment.core.application.port.in.assessmentinsight.ApproveAssessmentInsightUseCase;
+import org.flickit.assessment.core.application.port.in.insight.assessmentinsight.ApproveAssessmentInsightUseCase;
 import org.flickit.assessment.core.application.port.out.assessmentinsight.ApproveAssessmentInsightPort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
