+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/service/insight/assessmentinsight/GetAssessmentInsightService.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.application.service.assessmentinsight;
+package org.flickit.assessment.core.application.service.insight.assessmentinsight;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.application.domain.assessment.AssessmentAccessChecker;
@@ -7,7 +7,7 @@ import org.flickit.assessment.common.exception.AccessDeniedException;
 import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.core.application.domain.AssessmentInsight;
 import org.flickit.assessment.core.application.domain.AssessmentResult;
-import org.flickit.assessment.core.application.port.in.assessmentinsight.GetAssessmentInsightUseCase;
+import org.flickit.assessment.core.application.port.in.insight.assessmentinsight.GetAssessmentInsightUseCase;
 import org.flickit.assessment.core.application.port.out.assessmentinsight.LoadAssessmentInsightPort;
 import org.flickit.assessment.core.application.port.out.assessmentresult.LoadAssessmentResultPort;
 import org.springframework.stereotype.Service;
