+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/service/insight/assessmentinsight/CreateAssessmentInsightServiceTest.java
@@ -1,11 +1,11 @@
-package org.flickit.assessment.core.application.service.assessmentinsight;
+package org.flickit.assessment.core.application.service.insight.assessmentinsight;
 
 import org.flickit.assessment.common.application.domain.assessment.AssessmentAccessChecker;
 import org.flickit.assessment.common.exception.AccessDeniedException;
 import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.core.application.domain.AssessmentInsight;
-import org.flickit.assessment.core.application.port.in.assessmentinsight.CreateAssessmentInsightUseCase;
-import org.flickit.assessment.core.application.port.in.assessmentinsight.CreateAssessmentInsightUseCase.Param;
+import org.flickit.assessment.core.application.port.in.insight.assessmentinsight.CreateAssessmentInsightUseCase;
+import org.flickit.assessment.core.application.port.in.insight.assessmentinsight.CreateAssessmentInsightUseCase.Param;
 import org.flickit.assessment.core.application.port.out.assessmentinsight.CreateAssessmentInsightPort;
 import org.flickit.assessment.core.application.port.out.assessmentinsight.LoadAssessmentInsightPort;
 import org.flickit.assessment.core.application.port.out.assessmentinsight.UpdateAssessmentInsightPort;
